% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_demo.R
\name{build_demo}
\alias{build_demo}
\title{build a XenSCE instance for GBM demo data from TENx}
\usage{
build_demo()
}
\description{
build a XenSCE instance for GBM demo data from TENx
}
\note{
This can take some time to run, especially if no resources
have been cached yet.  The data are derived from
`https://www.10xgenomics.com/datasets/ffpe-human-brain-cancer-data-with-human-immuno-oncology-profiling-panel-and-custom-add-on-1-standard`.
}
\examples{
if (interactive()) {
  gbmdemo = build_demo()
  gbmdemo
}
}
