% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_panc_subset.R
\name{build_panc_subset}
\alias{build_panc_subset}
\title{build a XenSCE instance based on the SFEData XeniumOutput element}
\usage{
build_panc_subset(cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{cache}{BiocFileCache::BiocFileCache() is default}
}
\description{
build a XenSCE instance based on the SFEData XeniumOutput element
}
\note{
This builds a XenSCE instance with on-disk counts.  `data(panc_sub)`
retrieves an instance with counts in dgCMatrix format, in memory.
}
\examples{
panc_sub_h5 = build_panc_subset()
panc_sub_h5
}
