% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_assets.R
\name{cache_assets}
\alias{cache_assets}
\title{return paths to cached entities after retrieving and caching them if needed}
\usage{
cache_assets(cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{cache}{instance of BiocFileCache::BiocFileCache()}
}
\description{
return paths to cached entities after retrieving and caching them if needed
}
\note{
On first usage around 4GB of data, mostly parquet, will be downloaded
and added to cache.
}
