% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_assets.R
\name{cache_demo}
\alias{cache_demo}
\title{helper for caching a built demo XenSCE}
\usage{
cache_demo(src, cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{src}{a XenSCE instance}

\item{cache}{instance of BiocFileCache::BiocFileCache()}
}
\value{
zero if `gbm_xen_demo.rds` is already present
in cache; otherwise result of BiocFileCache::bfcadd().
}
\description{
helper for caching a built demo XenSCE
}
\note{
Cache will hold an RDS file gbm_xen_demo.rds.
}
