% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_panc_subset.R
\name{cache_panc_subset}
\alias{cache_panc_subset}
\title{cache and/or retrieve path to zipped pancreas subset (originally from SFEData)}
\usage{
cache_panc_subset(cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{cache}{BiocFileCache::BiocFileCache() is default}
}
\description{
cache and/or retrieve path to zipped pancreas subset (originally from SFEData)
}
\examples{
cache_panc_subset()
}
