% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellBoundaries.R
\name{clip_rect}
\alias{clip_rect}
\title{restrict XenSCE to cells with centroids in specified rectangle}
\usage{
clip_rect(xsce, xlim, ylim)
}
\arguments{
\item{xsce}{XenSCE instance}

\item{xlim}{numeric(2)}

\item{ylim}{numeric(2)}
}
\description{
restrict XenSCE to cells with centroids in specified rectangle
}
