% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenSCE_class.R
\name{getCellBoundaries,XenSCE-method}
\alias{getCellBoundaries,XenSCE-method}
\title{method for cell boundary extraction}
\usage{
\S4method{getCellBoundaries}{XenSCE}(x)
}
\arguments{
\item{x}{instance of XenSCE}
}
\description{
method for cell boundary extraction
}
