% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenSCE_class.R
\name{getNucleusBoundaries,XenSCE-method}
\alias{getNucleusBoundaries,XenSCE-method}
\title{method for nucleus boundary extraction}
\usage{
\S4method{getNucleusBoundaries}{XenSCE}(x)
}
\arguments{
\item{x}{instance of XenSCE}
}
\description{
method for nucleus boundary extraction
}
