% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenSCE_class.R
\name{getTranscripts,XenSCE-method}
\alias{getTranscripts,XenSCE-method}
\title{method for transcript extraction}
\usage{
\S4method{getTranscripts}{XenSCE}(x)
}
\arguments{
\item{x}{instance of XenSCE}
}
\description{
method for transcript extraction
}
