% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellBoundaries.R
\name{plotCellBoundaries}
\alias{plotCellBoundaries}
\title{render boundaries of cells with optional centroid positions and transcript positions}
\usage{
plotCellBoundaries(
  xsce,
  add_cent = TRUE,
  cent_col = "red",
  cent_cex = 0.2,
  add_tx = TRUE,
  tx_cex = 0.1
)
}
\arguments{
\item{xsce}{XenSCE instance}

\item{add_cent}{logical(1)}

\item{cent_col}{character(1) default to "red"}

\item{cent_cex}{numeric(1) default to 0.2}

\item{add_tx}{logical(1)}

\item{tx_cex}{numeric(1)}
}
\description{
render boundaries of cells with optional centroid positions and transcript positions
}
\examples{
dem = build_panc_subset()
plotCellBoundaries(clip_rect(dem, xlim=c(600,850), ylim=c(500,750)))
}
