% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipXenSPEP.R
\name{restoreZipXenSPEP}
\alias{restoreZipXenSPEP}
\title{use unzip, readRDS, and loadGeometry to restore a XenSPEP}
\usage{
restoreZipXenSPEP(zipf)
}
\arguments{
\item{zipf}{character(1) path to zip file created with `zipXenSPEP`}
}
\value{
instance of XenSPEP
}
\description{
use unzip, readRDS, and loadGeometry to restore a XenSPEP
}
