% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_assets.R
\name{retrieve_demo}
\alias{retrieve_demo}
\title{retrieve cached demo XenSCE}
\usage{
retrieve_demo(cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{cache}{instance of BiocFileCache::BiocFileCache()}
}
\description{
retrieve cached demo XenSCE
}
