% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe2xsce.R
\name{sfe2xsce}
\alias{sfe2xsce}
\title{convert SpatialFeatureExperiment to XenSCE}
\usage{
sfe2xsce(sfe, BPPARAM = BiocParallel::SerialParam())
}
\arguments{
\item{sfe}{SpatialFeatureExperiment instance}

\item{BPPARAM}{see BiocParallel}
}
\description{
convert SpatialFeatureExperiment to XenSCE
}
\examples{
example(cache_sfeLung)
nx = sfe2xsce(ans, BPPARAM=BiocParallel::MulticoreParam(8))
nx
}
