% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_xen.R
\name{show,XenSPEP-method}
\alias{show,XenSPEP-method}
\title{display aspects of XenSPEP}
\usage{
\S4method{show}{XenSPEP}(object)
}
\arguments{
\item{object}{instance of XenSCE}
}
\description{
display aspects of XenSPEP
}
