% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{panc_sub}
\alias{panc_sub}
\title{example XenSCE instance based on pancreas subset from SFEData XeniumOutput}
\format{
XenSCE
}
\usage{
panc_sub
}
\description{
example XenSCE instance based on pancreas subset from SFEData XeniumOutput
}
\examples{
data(panc_sub)
head(panc_sub)
}
\keyword{datasets}
